chainspec refactor
